pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "acanag"
version = "0.0.1"
description = "Active Anomaly Aggregation"
authors = [
    {name = "Benjamin Auder",email = "benjamin@auder.net"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "matplotlib",
  "modAL-python",
  "numpy",
  "scikit-learn",
  "tensorflow",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
license = "MIT"
license-files = ["LICEN[CS]E*"]

[project.urls]
Homepage = "https://github.com/yagu0/ActiveAnomalyAggregation"
Issues = "https://github.com/yagu0/ActiveAnomalyAggregation/issues"
